Lightweight events tests moved in separate class.
