+++ targets/unity-v2/Testing/Tests/Client/ClientApiTests.cs
@@ -1,16 +1,9 @@
 #if !DISABLE_PLAYFABCLIENT_API
-using PlayFab.ClientModels;
-using PlayFab.Json;
 using System;
 using System.Collections.Generic;
+using PlayFab.ClientModels;
 using PlayFab.Internal;
-#if NET_4_6
-using System.Linq;
-using System.Threading.Tasks;
-using PlayFab.Logger;
-using PlayFab.Pipeline;
-using PlayFab.SharedModels;
-#endif
+using PlayFab.Json;
 
 namespace PlayFab.UUnit
 {
@@ -472,158 +465,6 @@ namespace PlayFab.UUnit
             // There's nothing else useful to test about this right now
             ((UUnitTestContext)result.CustomData).EndTest(UUnitFinishState.PASSED, null);
         }
-#if NET_4_6
-        [UUnitTest]
-        public void WriteOneDSEvents(UUnitTestContext testContext) => WriteOneDSEventsAsync(testContext);
-        /// <summary>
-        /// ONEDS API
-        /// Test that a batch of custom events can be sent to OneDS server
-        /// </summary>
-        public async void WriteOneDSEventsAsync(UUnitTestContext testContext)
-        {
-            var event1 = new PlayFabEvent() { Name = "Event_1", EventType = PlayFabEventType.Lightweight };
-            event1.SetProperty("Prop-A", true);
-            event1.SetProperty("Prop-B", "hello");
-            event1.SetProperty("Prop-C", 123);
-
-            var event2 = new PlayFabEvent() { Name = "Event_2", EventType = PlayFabEventType.Lightweight };
-            event2.SetProperty("Prop-A", false);
-            event2.SetProperty("Prop-B", "good-bye");
-            event2.SetProperty("Prop-C", 456);
-
-            var request = new EventsModels.WriteEventsRequest
-            {
-                Events = new List<EventsModels.EventContents>
-                {
-                    event1.EventContents,
-                    event2.EventContents
-                }
-            };
-
-            var oneDSEventsApi = new OneDSEventsAPI();
-
-            // get OneDS authentication from PlayFab
-            var configRequest = new EventsModels.TelemetryIngestionConfigRequest();
-            var authTask = await OneDSEventsAPI.GetTelemetryIngestionConfigAsync(configRequest);
-            var response = authTask.Result;
-            if (response != null)
-            {
-                oneDSEventsApi.SetCredentials("o:" + response.TenantId, response.IngestionKey, response.TelemetryJwtToken, response.TelemetryJwtHeaderKey, response.TelemetryJwtHeaderPrefix);
-            }
-            else
-            {
-                testContext.Fail("Failed to get OneDS authentication info from PlayFab");
-                return;
-            }
-
-            // call OneDS events API
-            var writeTask = oneDSEventsApi.WriteTelemetryEventsAsync(request, null, new Dictionary<string, string>());
-            ContinueWithContext(writeTask, testContext, WriteOneDSEventsContinued, true, "Failed to send a batch of custom OneDS events", true);
-        }
-        
-        private void WriteOneDSEventsContinued(PlayFabResult<EventsModels.WriteEventsResponse> writeResult, UUnitTestContext testContext, string failMessage)
-        {
-            testContext.True(writeResult.Error == null && writeResult.Result != null, "Failed to send a batch of custom events to OneDS server");
-        }
-
-        [UUnitTest]
-        public void EmitLightweightEvents(UUnitTestContext testContext) => EmitLightweightEventsAsync(testContext);
-        public async void EmitLightweightEventsAsync(UUnitTestContext testContext)
-        {
-            // create and set settings for OneDS event pipeline
-            var settings = new OneDSEventPipelineSettings();
-            settings.BatchSize = 8;
-            settings.BatchFillTimeout = TimeSpan.FromSeconds(1);
-
-            var logger = new DebugLogger();
-
-            // create OneDS event pipeline
-            var oneDSPipeline = new OneDSEventPipeline(settings, logger);
-
-            // create custom event API, add the pipeline
-            var playFabEventApi = new PlayFabEventAPI(logger);
-            playFabEventApi.EventRouter.AddAndStartPipeline(EventPipelineKey.OneDS, oneDSPipeline);
-
-            // create and emit many lightweight events
-            var results = new List<Task>();
-            for (int i = 0; i < 50; i++)
-            {
-                results.AddRange(playFabEventApi.EmitEvent(CreateSamplePlayFabEvent("Event_Custom", PlayFabEventType.Lightweight)));
-            }
-
-            // wait when the pipeline finishes sending all events
-            await Task.WhenAll(results);
-
-            // check results
-            var sentBatches = new Dictionary<IList<IPlayFabEmitEventRequest>, int>();
-            foreach (var result in results)
-            {
-                testContext.True(result.IsCompleted, "Custom event emission task failed to complete");
-                PlayFabEmitEventResponse response = (PlayFabEmitEventResponse)((Task<IPlayFabEmitEventResponse>)result).Result;
-                testContext.True(response.EmitEventResult == EmitEventResult.Success, "Custom event emission task failed to succeed");
-                testContext.True(response.PlayFabError == null && response.WriteEventsResponse != null, "Custom event failed to be sent");
-
-                if (!sentBatches.ContainsKey(response.Batch))
-                    sentBatches[response.Batch] = 0;
-                sentBatches[response.Batch]++;
-            }
-
-            // 6 full batches of 8 events and 1 incomplete batch of 2 events are expected
-            testContext.True(sentBatches.Count(x => x.Value == 8) == 6, "Wrong number of full batches");
-            testContext.True(sentBatches.Count(x => x.Value == 2) == 1, "Wrong number of incomplete batches");
-            testContext.EndTest(UUnitFinishState.PASSED, null);
-        }
-
-        private PlayFabEvent CreateSamplePlayFabEvent(string name, PlayFabEventType eventType)
-        {
-            var customEvent = new PlayFabEvent { Name = name, EventType = eventType };
-            customEvent.SetProperty("Prop-A", true);
-            customEvent.SetProperty("Prop-B", "custom");
-            customEvent.SetProperty("Prop-C", 567);
-            return customEvent;
-        }
-
-        private static void ContinueWithContext<T>(Task<PlayFabResult<T>> srcTask, UUnitTestContext testContext,
-            Action<PlayFabResult<T>, UUnitTestContext, string> continueAction, bool expectSuccess, string failMessage,
-            bool endTest) where T : PlayFabResultCommon
-        {
-            srcTask.ContinueWith(task =>
-                {
-                    var failed = true;
-                    try
-                    {
-                        if (expectSuccess)
-                        {
-                            testContext.NotNull(task.Result, failMessage);
-                            testContext.IsNull(task.Result.Error);
-                            testContext.NotNull(task.Result.Result, failMessage);
-                        }
-
-                        continueAction?.Invoke(task.Result, testContext, failMessage);
-                        failed = false;
-                    }
-                    catch (UUnitSkipException uu)
-                    {
-                        // Silence the assert and ensure the test is marked as complete - The exception is just to halt the test process
-                        testContext.EndTest(UUnitFinishState.SKIPPED, uu.Message);
-                    }
-                    catch (UUnitException uu)
-                    {
-                        // Silence the assert and ensure the test is marked as complete - The exception is just to halt the test process
-                        testContext.EndTest(UUnitFinishState.FAILED, uu.Message + "\n" + uu.StackTrace);
-                    }
-                    catch (Exception e)
-                    {
-                        // Report this exception as an unhandled failure in the test
-                        testContext.EndTest(UUnitFinishState.FAILED, e.ToString());
-                    }
-
-                    if (!failed && endTest)
-                        testContext.EndTest(UUnitFinishState.PASSED, null);
-                }
-            );
-        }
-#endif
     }
 }
 #endif
